allow app context to be available during background tasks by breaking import cycle
