+++ services/web/backend/flask_app/__init__.py
@@ -1,0 +1,9 @@
+from . import app
+from . import db
+from . import emails
+from . import modeling
+from . import settings
+from . import utils
+from . import version
+
+__all__ = ["app", "emails", "settings", "utils", "db", "version", "modeling"]
